requirements.txt
alpha_vantage
requests
openai
pandas
python-dotenv
matplotlib
seaborn
scipy 